refactor: split add command workflow
Keep cmd/malt/add.go focused on Cobra wiring and output formatting while moving add options, staging, materialization, tree helpers, workflow dispatch, and MerkleDAG import code into focused files.

Add a source guard that prevents add.go from regressing into layout/materialization logic.
